backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-dotenv
httpx
minio
celery
redis
python-jose
email-validator
passlib
python-multipart
